Notifier/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
arrow==1.2.2
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.11.1
blinker==1.4
certifi==2022.5.18.1
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
distro==1.7.0
docopt==0.6.2
greenlet==1.1.2
httplib2==0.20.2
icalendar==4.0.9
ics==0.7.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
macaroonbakery==1.3.1
more-itertools==8.10.0
msgpack==1.0.3
multidict==6.0.2
netifaces==0.11.0
numpy==1.23.1
oauthlib==3.2.0
olefile==0.46
pandas==1.4.3
pexpect==4.8.0
Pillow==9.0.1
pipreqs==0.4.11
protobuf==3.12.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyaes==1.6.1
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.21
PyJWT==2.3.0
pylibacl==0.6.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
python-dateutil==2.8.1
python-dotenv==0.20.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.0
rsa==4.9
schedule==1.1.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.2.post1
systemd-python==234
TatSu==5.8.0
telepot==12.7
Telethon==1.24.0
tornado==6.1
urllib3==1.26.9
wadllib==1.3.6
xdg==5
yarg==0.1.9
yarl==1.7.2
zipp==1.0.0
